CRUD tienda

+++ src/main/java/com/tienda/tiendita/repository/ProveedoresRepository.java
@@ -4,5 +4,5 @@ import com.tienda.tiendita.models.ProductosModel;
 import com.tienda.tiendita.models.ProveedoresModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProveedoresRepository extends JpaRepository<ProveedoresModel, Long> {
+public interface ProveedoresRepository extends JpaRepository<ProveedoresModel, Integer> {
 }
