+++ src/main/java/com/tienda/tiendita/services/interfaces/IproveedoresService.java
@@ -10,12 +10,12 @@ public interface IproveedoresService {
 
     public List<ProveedoresModel> listaProveedores();
 
-    public ProveedoresModel buscarProveedorePorId(Long idProveedor);
+    public ProveedoresModel buscarProveedorePorId(Integer idProveedor);
 
     public ResponseEntity<RespuestaApi> guardarProveedor(ProveedoresModel proveedor);
 
-    public ResponseEntity<RespuestaApi> eliminarProveedor(Long id);
+    public ResponseEntity<RespuestaApi> eliminarProveedor(Integer id);
 
-    public ResponseEntity<RespuestaApi> actualizarProducto(Long id, ProveedoresModel proveedor);
+    public ResponseEntity<RespuestaApi> actualizarProducto(Integer id, ProveedoresModel proveedor);
 
 }
